requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
pinecone-client==2.2.4
pdfplumber==0.9.0
transformers==4.35.0
torch==2.2.0
requests==2.31.0
python-multipart==0.0.6